Moved MockServer to gRPCMockClasses
